Recursion/Sort fixups

+++ programs/recursive-functions/flood-fill-start.cpp
@@ -3,10 +3,10 @@
 using namespace std;
 
 void printMaze(const vector<vector<char>>& curMaze) {
-    int rows = curMaze.size();
-    int cols = curMaze.at(0).size();
-    for (int r = 0; r < rows; r++) {
-        for (int c = 0; c < cols; c++) {
+    size_t rows = curMaze.size();
+    size_t cols = curMaze.at(0).size();
+    for (size_t r = 0; r < rows; r++) {
+        for (size_t c = 0; c < cols; c++) {
             cout << curMaze.at(r).at(c);
         }
         cout << endl;
